+++ packages/common/src/types/api/file.ts
@@ -1,0 +1,18 @@
+export interface FileUploadData {
+  file: File;
+  userId: string;
+}
+
+export interface HeadshotUploadData extends FileUploadData {
+  fullName: string;
+}
+
+export interface UploadResult {
+  success: true;
+  key: string;
+}
+
+export interface UploadError {
+  success: false;
+  error: string;
+}
