+++ packages/common/src/types/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./application";
+export * from "./file";
