+++ packages/common/src/types/entities/file.ts
@@ -1,12 +1,3 @@
-export interface FileUploadData {
-  file: File;
-  userId: string;
-}
-
-export interface HeadshotUploadData extends FileUploadData {
-  fullName: string;
-}
-
 export interface FileValidationConfig {
   maxBytes: number;
   allowedMimeTypes: Set<string>;
